Extract intcode
